Rework hooks
Hooks now back up the context (registers and global instances) making recursive hooks possible. A buffer is added to the data stack and the stack pointer is remembered before calling the deadalus functions to handle wrongful popping or remaining data left on the data stack. Additionally, the deadalus functions are now called with an event handler. This applies the new safety measures in-between each function call and removes the need for the typical "once"-hook initialization, turning the hook-initialization into an automatic hook-once-initialization.
For that the event handler was minimally extended to also allow pointers instead of handles.
As hooks are fundamental to many LeGo packages, this revision should be used with caution until stability is confirmed (so far tested with Gothic 1 and Gothic 2, with recursive hooks, and with functions that change global instances).

git-svn-id: file:///home/lehona/Downloads/lego_svn/lego@139 2e52861e-7267-437c-9073-90786e7a2ebf