Allow direct and indirect Daedalus hooks
Previously, Daedalus hooks replaced the pointer to the position in the
code stack of the hooked function symbol (symbol.content). This had the
disadvantage of only allowing each function to be hooked once.

More importantly, this only worked with indirect function calls, where the
target position in the code stack was resolved at run time with the
function's symbol, e.g. MEM_Call(funcname), but not when the call had been
reduced to a jump in the code stack to a fixed position, e.g. funcname();

Instead of replacing the code stack pointer of the symbol, the byte code
at the respective position is now replaced with a jump to the hook
function. The overwritten byte code is then executed on ContinueCall();
before jumping back to the next position in the code stack. This is very
similar to how the engine hooks were originally designed and thus allows
multiple hooks in the same function. However, a function can only be used
once to hook, because it's symbol index can only be stored once in the
hash table.

Side note: Functions can also hook AFTER Daedalus functions by first
passing all arguments, continuing the call and doing new things
afterwards, as the hooked function returns into the hook function.