Get flags in humand-readable format
A new function returns the names of the LeGo flags contained in a provided
bitfield. Its purpose is to offer insight into the initialized LeGo
packages. To aid in that, a console command is provided (only registered
if LeGo is initialized with LeGo_ConsoleCommands). When typing 'LeGo' into
the game console the current LeGo version as well as the initialized
packages are returned as output. This information is also sent to the zSpy
on initialization.