Fix FF without delay if PiM or GT
So far FF have always been called every frame if their delay was zero,
irrespective if the timer was paused in the menu. This was due to a
condition that was always satisfied as long as the timer did not
increase (timer >= itm.next).

The issue is now circumvented by increasing itm.next minimally (one
millisecond) if itm.delay is zero. Since a frame should always last at
least one millisecond, this should not be a problem (fps < 1000).