Add safety checks for return value of hook
To not jeopardize existing hooks that might accidentally have a non-void
return value and have left data on the stack, the token on top of the
stack is checked against zPAR_TOK_PUSHINT (present when returning an
integer or float value or variable). This offers no guarantee, but
increased probability that the value on top of the stack was intended.