Support pushed variables as hook return values
Caution!

Before using the changes of the previous commits, it is important to
double-check all present engine hooks.

The return type of all Daedalus functions that hook the engine should be
void, if not explicitly intended otherwise. Since return values are only
supported from here on out, this means, that the user has to confirm
that all prior existing engine hooks [1] are of the form

    func void

Incorrect return types are no longer gracefully ignored but may lead to
severe and hard to track undefined behavior or crashes.

[1] An engine hook is any function referenced in a function call of
        HookEngine, HookEngineF, HookEngineI, HookEngineS
    It is a simple file-search operation to find these engine hooks and
    to correct their function return type to void.