Improve performance of saving bound strings
Saving a string that is bound to a handle (i.e. PM_BindString) requires
finding the symbol of that string. Instead of iterating over the entire
symbold table for each bound string, as done before, all string symbols
are now only collected once. This reduces archiving time of PermMem by
up to 200 ms for each bound string (depending on the total number of
symbols).