Prevent problems with Locals and repeat
The Ikarus function repeat cannot be safely used in combination with
Locals, because nested calls will overwrite the loop condition. Instead
while loops are used where applicable. Partially, this was already the
case (foreachHndl), but was problematic in HookEngine and Hashtable
(foreach).
Additionally, the foreach-function of arrays introduced recently was
fixed.