Fix handles slipping through PM reset
The changes around 97e5e caused some handles not to be properly deleted
when resetting PermMem. With the hash bucket being manipulated at every
iteration, the number of iterations changed and some handles were
skipped. Now the bucket is properly copied as done already in other loop
functions of arrays. The same change as been applied to the HookEngine
functions.

Ref szapp/Ninja#32